requirements.txt
langgraph 
langchain 
langchain-community 
python-dotenv 
langchain-ollama
requests
pydantic